Scale Tweaks and Product Weight
Added weight stability checking in scale.js, since the USB scale was inaccurate in depicting changing weight. Typing in a Product Lookup Code (PLU) in the "Type in SKU" overlay will correctly pull in an item. Any item with a weight price instead of unit price will be required to be placed on the scale.